requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
pypdf
faiss-cpu
pyngrok
google-generativeai
streamlit